requirements.txt
requests>=2.9
lxml>=3.5
eventlet>=0.18
ebooklib>=0.15
configargparse>=0.10
pyyaml>=3
